set correct `iPoolId` for miner_work
- initialize miner_work iPoolId in the default constructor with the invalid pool id
- move definition of `pool_data` into an own file
